Serperate page allocator code from startup's one
